pyproject.toml
[project]
name = "lifgame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "gunicorn>=23.0.0",
    "line-bot-sdk>=3.0",
    "openai>=2.14.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
